Config - Convertsion of r_model from double to float added Constants sres changed to float[] from double[] ILGPU updated to 0.10.1 Program removed testing of spline code Spectral_model - match spectral resolution mostly translated - Accuracy tests to follow Spectral_resolution - Added interpolator Spectral_resolution - Added Methods : match, GaussianKernelDifference, DetermineVariance, convert_vd2pd, finalize_GaussianKernelDifference, adjust_resolution, convert_pd2vd Utility Methods : Added Methods : Diff, Scale, WhereIsClose, WhereNot Vector - Finalised changes to the Fill method simplification, Added Method Power, PowerAbs, Simplified Dot Product Method

+++ Constants.cs
@@ -112,7 +112,7 @@ namespace MachineLearningSpectralFittingCode
         //
         public static float slope { get; set; } = 0f;
         public static int sidx { get; set; } = 0;
-        public static double[] sres { get; set; }
+        public static float[] sres { get; set; }
 
         // Dust Maps
         public static float[,] ngp_dust { get; set; }
